requirements.txt
django
django-registration==0.8
pillow
South
django-extensions
django-debug-toolbar
django-grappelli
django-sitetree
ipython
mysql-python
fabric
django-bootstrap-form
django-reversion
django-markitup
postmarkup
django-tagging
